fix: ios step def n debug mode
