+++ features/scenarios/non-xray/cucumber/apps/ios/scenarios.feature
@@ -3,4 +3,4 @@ Feature: Automation iOS
 
   @ios-tc-01
   Scenario: User access action bar tab on iOS
-    When user open iOS apps
+    Given user open iOS apps
