--- features/stepdefinitions/apps/ios/api_demos_steps.go
@@ -1,11 +1,0 @@
-package stepdefinitions
-
-import (
-	"fmt"
-)
-
-func OpenIOS() error {
-	fmt.Println("aaa")
-
-	return nil
-}
