+++ features/suites/apps/ios/suites.go
@@ -1,9 +1,9 @@
 package suites
 
 import (
-	stepdefinitions "github.com/golang-automation-v1/features/stepdefinitions/apps/ios"
-
 	"github.com/cucumber/godog"
+
+	stepdefinitions "github.com/golang-automation-v1/features/stepdefinitions/apps/ios"
 )
 
 func IOSScenarioContext(s *godog.ScenarioContext) {
