+++ features/supports/drivers/web_driver.go
@@ -18,14 +18,11 @@ func webProperties() *properties.Properties {
 func WebCapabilities() {
 	var err error
 
-	fmt.Println("1")
 	p := webProperties()
-	fmt.Println("2")
 
 	caps := selenium.Capabilities{
 		"browserName": p.MustGetString("browserName"),
 	}
-	fmt.Println("3")
 
 	WebDriver, err = selenium.NewRemote(caps, "")
 	fmt.Println(err)
